itsm_backend/pyproject.toml
[tool.poetry]
name = "itsm-project"
version = "0.1.0"
description = ""
authors = ["Corner324 <mosenn@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
djangorestframework-simplejwt = "^5.3.1"
django-debug-toolbar = "^4.4.6"
django-admin-interface = "^0.29.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
